requirements.txt
Django==1.7.1
Fabric==1.4.3
MarkupSafe==0.23
PyYAML==3.11
bcdoc==0.12.0
boto==2.34.0
botocore==0.31.0
clan==0.2.3
colorama==0.2.5
cssmin==0.1.4
django-grappelli==2.6.3
django-sortedm2m==0.8.1
docutils==0.11
facebook-sdk==0.4.0
google-api-python-client==1.3.1
gunicorn==19.1.1
httplib2==0.9
jmespath==0.2.1
nose==1.2.1
odict==1.5.1
openpyxl==1.8.5
ply==3.4
psycopg2==2.5.4
pyasn1==0.1.7
pycrypto==2.6
python-dateutil==2.2
requests==2.4.3
rsa==3.1.2
six==1.5.2
slimit==0.7.4
smartypants==1.8.6
ssh==1.7.14
termcolor==1.1.0
twitter==1.14.3
wheel==0.23.0
wsgiref==0.1.2
